fix minor issues
