+++ .golangci.yaml
@@ -1,0 +1,34 @@
+linters:
+  enable:
+    - asciicheck
+    - dogsled
+    - dupl
+    - errcheck
+    - errorlint
+    - exportloopref
+    - gocognit
+    - goconst
+    - gocyclo
+    - gofmt
+    - goimports
+    - gosec
+    - gosimple
+    - govet
+    - ineffassign
+    - misspell
+    - nakedret
+    - prealloc
+    - revive
+    - staticcheck
+    - stylecheck
+    - unconvert
+    - unused
+
+linters-settings:
+  goimports:
+    local-prefixes: consensus-shipyard/go-ipc-types
+  gocognit:
+    min-complexity: 50
+
+run:
+  timeout: 5m
