+++ gen/gen.go
@@ -1,10 +1,11 @@
 package main
 
 import (
+	gen "github.com/whyrusleeping/cbor-gen"
+
 	"github.com/consensus-shipyard/go-ipc-types/gateway"
 	"github.com/consensus-shipyard/go-ipc-types/sdk"
 	"github.com/consensus-shipyard/go-ipc-types/subnetactor"
-	gen "github.com/whyrusleeping/cbor-gen"
 )
 
 func main() {
